requirements.txt
matplotlib
pandas
streamlit
Pillow
